refactor(handoff): de-duplicate publisher modifier, split lookup protocols
Code-review on commit f1d4356d:
- Extract HandoffPublisherModifier to its own file (was byte-identical
  duplicate across ProfileWindowView.swift and ProfileRootView.swift).
- Split HandoffAccountLookup and HandoffEarmarkLookup into their own
  files (one primary type per file, CODE_GUIDE §2).
- Parameterise the staticTitles test via @Test(arguments:) so each
  destination shows as its own row in the Test Navigator.
- Add docstring to HandoffTitleProvider.title(for:accounts:earmarks:)
  describing fallback behaviour and the @MainActor rationale.
