+++ App/ProfileWindowView.swift
@@ -174,46 +174,4 @@
 
   }
 
-  /// Reads the focused window's navigation state (sidebar / selected
-  /// transaction / analysis / reports) and stamps an `NSUserActivity` via
-  /// SwiftUI's `.userActivity(_:isActive:_:)` so the system can hand the
-  /// route off to a peer device. Attached inside the `.ready` session
-  /// branch so the `AccountStore` / `EarmarkStore` environment values
-  /// (read for title lookup) are guaranteed to be present.
-  private struct HandoffPublisherModifier: ViewModifier {
-    let profileID: UUID
-    @Environment(AccountStore.self) private var accountStore
-    @Environment(EarmarkStore.self) private var earmarkStore
-    @FocusedValue(\.sidebarSelection) private var sidebarSelection
-    @FocusedValue(\.selectedTransactionID) private var selectedTransactionID
-    @FocusedValue(\.analysisRoute) private var analysisRoute
-    @FocusedValue(\.reportsRoute) private var reportsRoute
-
-    private var route: NavigationDestination? {
-      // `sidebarSelection` is `Binding<SidebarSelection?>?`. Flatten via
-      // `flatMap` on the outer optional to a plain `SidebarSelection?`.
-      let sidebar = sidebarSelection.flatMap { $0.wrappedValue }
-      return NavigationDestination.make(
-        sidebar: sidebar,
-        selectedTransaction: selectedTransactionID,
-        analysis: analysisRoute,
-        reports: reportsRoute)
-    }
-
-    func body(content: Content) -> some View {
-      content.userActivity(
-        HandoffActivity.continueActivityType,
-        isActive: route != nil
-      ) { activity in
-        guard let route else { return }
-        let payload = HandoffPayload(profileID: profileID, destination: route)
-        let title = HandoffTitleProvider.title(
-          for: route,
-          accounts: accountStore,
-          earmarks: earmarkStore)
-        activity.configureHandoff(payload: payload, title: title)
-      }
-    }
-  }
-
 #endif
