+++ Automation/Handoff/HandoffAccountLookup.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+/// Read-only window into account display names. Implemented by the real
+/// `AccountStore` and by fakes in tests.
+@MainActor
+protocol HandoffAccountLookup {
+  func displayName(for id: UUID) -> String?
+}
